Cleaned up libraries and changes the high level modules include statements
made requested changes from pull request

fixed typo and missed revision

removing unit change

final changes to libraries
