+++ libraries/Settings/SettingsTemplate.h
@@ -1,5 +1,5 @@
-#ifndef _SITE_TRICYCLE_SETTINGS_HEADER_
-#define _SITE_TRICYCLE_SETTINGS_HEADER_
+#ifndef _SITE_TRICYCLE_SETTINGS_HEADER_H
+#define _SITE_TRICYCLE_SETTINGS_HEADER_H
 
 /*
 Settings.h is a site-specific set of trike parameters.
@@ -77,6 +77,20 @@ VEHICLE_NUMBER to your own new number.
 // DISK_BRAKE is deprecated, use BRAKE_OUT_PIN
 // Use DAC A
 #define THROTTLE_CHANNEL 0
+
+// Defines from the old common.h
+#define WHEEL_BASE_mm         800
+#define TURNING_RADIUS_mm    4000
+#define MAX_ACCEL_mmPs2      2000
+#define STANDARD_ACCEL_mmPs2 1000
+// but we send a number from 0 to 255
+#define STANDARD_ACCEL       55
+#define MAX_ACCEL            110
+#define HALF_BRAKE           127
+#define FULL_BRAKE           255
+#define WALK_SPEED_mmPs      1000
+// 10 mph = 4.44 m/s
+#define MAX_SPEED_mmPs       4444
 
 // Parameters used by RC_Control_interrupts
 // Indices for information stored in array RC_results
